[RenderTreeBuilder] Do not try to collapse anonymous blocks when cleaning up the first-letter subtree https://bugs.webkit.org/show_bug.cgi?id=224667 <rdar://76616098>
Reviewed by Antti Koivisto.

We construct a dedicated renderer for the first letter content and in some cases wrap it into an anonymous block.
This subtree needs to be destroyed when the associated RenderText is getting deleted. During this cleanup process
we may find the wrapper anonymous block redundant and initiate a sibling block collapsing. However this collapsing
may interfere with the initial delete (the associated RenderText) and lead to unexpected result.
Let's reserve the anonymous collapsing for the initial remove.

* rendering/updating/RenderTreeBuilderFirstLetter.cpp:
(WebCore::RenderTreeBuilder::FirstLetter::cleanupOnDestroy):


Canonical link: https://commits.webkit.org/236631@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276137 268f45cc-cd09-0410-ab3c-d52691b4dbfc