Blend using calc() when necessary https://bugs.webkit.org/show_bug.cgi?id=224670
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

Mark 141 WPT progressions.

* web-platform-tests/css/css-backgrounds/animations/background-position-interpolation-expected.txt:
* web-platform-tests/css/css-backgrounds/animations/background-position-origin-interpolation-expected.txt:
* web-platform-tests/css/css-backgrounds/animations/border-image-width-interpolation-expected.txt:
* web-platform-tests/css/css-backgrounds/animations/border-radius-interpolation-expected.txt:
* web-platform-tests/css/css-flexbox/animation/flex-basis-composition-expected.txt:
* web-platform-tests/css/css-position/animations/bottom-composition-expected.txt:
* web-platform-tests/css/css-position/animations/left-composition-expected.txt:
* web-platform-tests/css/css-position/animations/right-composition-expected.txt:
* web-platform-tests/css/css-position/animations/top-composition-expected.txt:
* web-platform-tests/css/css-shapes/animation/shape-margin-composition-expected.txt:
* web-platform-tests/css/css-sizing/animation/height-composition-expected.txt:
* web-platform-tests/css/css-sizing/animation/max-height-composition-expected.txt:
* web-platform-tests/css/css-sizing/animation/max-width-composition-expected.txt:
* web-platform-tests/css/css-sizing/animation/min-height-composition-expected.txt:
* web-platform-tests/css/css-sizing/animation/min-width-composition-expected.txt:
* web-platform-tests/css/css-transitions/animations/vertical-align-composition-expected.txt:
* web-platform-tests/css/css-transitions/animations/vertical-align-interpolation-expected.txt:
* web-platform-tests/css/css-values/animations/calc-interpolation-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/interpolation-per-property-001-expected.txt:

Source/WebCore:

We need to blend between lengths when their types are not the same. However, we would take
some shortcuts in the following cases and blend using just one value:

1. either the from or to value was a 0 value,
2. progress was 0 or 1.

Now we use the mixed types code path if the types are different or either the from or to
value is a calculated value. The single exception where we don't actually end up generating
a CalculationValue is when we know that one of the two values is 0 and not a percentage, in
which case we run a simple blend.

Due to this approach, we stopped returning the value as-is in some cases where progress was
0 or 1. To avoid regressions in this area, we made the LengthBoxPropertyWrapper clone the from
or to value in the discrete case, as some other areas of the code do not know how to deal with
calculated Length values, for instance valueForImageSliceSide(const Length&).

* animation/CSSPropertyAnimation.cpp:
* platform/Length.cpp:
(WebCore::blendMixedTypes):
(WebCore::blend):
* platform/LengthBox.h:
(WebCore::LengthBox::LengthBox):
* platform/RectEdges.h:
(WebCore::RectEdges::RectEdges):

Canonical link: https://commits.webkit.org/236632@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276138 268f45cc-cd09-0410-ab3c-d52691b4dbfc