Web Inspector: Grid overlay column line label placement copy/paste error https://bugs.webkit.org/show_bug.cgi?id=224777
Reviewed by Devin Rousso.

In grid overlays, a typo meant for all column line labels after the initial line, only the start of the line was
used for placement, where it is more accurate to use the start and end points to match the behavior of the first
line. This had no visual impact because subsequent lines do not currently have their labels adjusted further into
the grid, only flipped into the grid but still referencing the same point. This patch corrects the value of
`gapLabelLine` to always be a line on which the label can be placed, instead of a single point.

* inspector/InspectorOverlay.cpp:
(WebCore::InspectorOverlay::buildGridOverlay):


Canonical link: https://commits.webkit.org/236756@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276274 268f45cc-cd09-0410-ab3c-d52691b4dbfc