REGRESSION (r276189): GPUProcess.WebProcessTerminationAfterTooManyGPUProcessCrashes is crashing https://bugs.webkit.org/show_bug.cgi?id=224790 <rdar://problem/76869318>
Reviewed by Darin Adler.

Source/WebKit:

GPUProcess.WebProcessTerminationAfterTooManyGPUProcessCrashes is intentionally repeatedly
killing the GPUProcess. As a result, the GPUProcess may get killed very shortly after
a relaunch and the RemoteAudioDestinationManager::StartAudioDestination synchronous IPC
may fail if it is ongoing at the time of the crash. This would cause m_destinationID to
not get initialized and then get sent as IPC parameter, thus crashing.

pre-r276189, we were not crashing because m_destinationID was not reset on crash and we
would thus send IPC for a destination that does not exist but at least the destinationID
would not be 0. This patch makes sure we don't try and send IPC at all when m_destinationID
is 0.

* WebProcess/GPU/media/RemoteAudioDestinationProxy.cpp:
(WebKit::RemoteAudioDestinationProxy::connection):
(WebKit::RemoteAudioDestinationProxy::startRendering):
(WebKit::RemoteAudioDestinationProxy::stopRendering):
(WebKit::RemoteAudioDestinationProxy::storageChanged):
* WebProcess/GPU/media/RemoteAudioDestinationProxy.h:

Tools:

Re-enable API test now that it is no longer crashing.

* TestWebKitAPI/Tests/WebKitCocoa/GPUProcess.mm:
(TEST):


Canonical link: https://commits.webkit.org/236766@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276284 268f45cc-cd09-0410-ab3c-d52691b4dbfc