[GPUP] Refactor SourceBuffer::enqueuedSamplesForTrackID() for testing purposes https://bugs.webkit.org/show_bug.cgi?id=224891
Reviewed by Eric Carlson.

Source/WebCore:

This patch is a preparation to run `MockMediaPlayerMediaSource` in the GPU Process.

`SourceBuffer::enqueuedSamplesForTrackID()` is a synchronous function used for testing
purposes. It works fine when "Media in GPU process" is disabled. However, when
"Media in GPU process" is enabled, we have to use a synchronous IPC message to
implement it. This patch changes it to be a function with a completion handler
parameter, so that we can implement it with an asynchronous IPC message when
we run `MockMediaPlayerMediaSource` in the GPU process.

There is no behavior change when "Media in GPU Process" is disabled.

Covered by existing tests.

* Modules/mediasource/SourceBuffer.cpp:
(WebCore::SourceBuffer::enqueuedSamplesForTrackID):
* Modules/mediasource/SourceBuffer.h:
* platform/graphics/SourceBufferPrivate.cpp:
(WebCore::SourceBufferPrivate::enqueuedSamplesForTrackID):
* platform/graphics/SourceBufferPrivate.h:
(WebCore::SourceBufferPrivate::enqueuedSamplesForTrackID): Deleted.
* platform/mock/mediasource/MockSourceBufferPrivate.cpp:
(WebCore::MockSourceBufferPrivate::enqueuedSamplesForTrackID):
* platform/mock/mediasource/MockSourceBufferPrivate.h:

* testing/Internals.cpp:
(WebCore::Internals::enqueuedSamplesForTrackID):
* testing/Internals.h:
* testing/Internals.idl:

Source/WebKit:

* GPUProcess/media/RemoteSourceBufferProxy.cpp:
(WebKit::RemoteSourceBufferProxy::enqueuedSamplesForTrackID):
* GPUProcess/media/RemoteSourceBufferProxy.h:
* GPUProcess/media/RemoteSourceBufferProxy.messages.in:
* WebProcess/GPU/media/SourceBufferPrivateRemote.cpp:
(WebKit::SourceBufferPrivateRemote::bufferedSamplesForTrackId):
(WebKit::SourceBufferPrivateRemote::enqueuedSamplesForTrackID):
* WebProcess/GPU/media/SourceBufferPrivateRemote.h:

LayoutTests:

Update following tests to use the asynchronous `enqueuedSamplesForTrackID()`.
* media/media-source/media-source-append-acb-no-frame-lost.html:
* media/media-source/media-source-append-acb-tolerance.html:
* media/media-source/media-source-append-presentation-durations.html:
* media/media-source/media-source-dropped-iframe.html:
* media/media-source/media-source-erase-after-last-append.html:
* media/media-source/media-source-first-append-not-starting-at-zero.html:
* media/media-source/media-source-remove-b-frame.html:
* media/media-source/media-source-seek-back.html:
* media/media-source/media-source-timestampoffset-rounding-error.html:

We don't need this platform-specific test expectation now.
* platform/mac/media/media-source/media-source-append-before-last-range-no-quota-exceeded-expected.txt: Removed.


Canonical link: https://commits.webkit.org/236879@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276414 268f45cc-cd09-0410-ab3c-d52691b4dbfc