[iOS] GPU Process sandbox lacks IOMobileFramebufferUserClient method filter https://bugs.webkit.org/show_bug.cgi?id=224956 <rdar://problem/68227590>
Reviewed by Per Arne Vollan.

This patch updates the GPU process sandbox to match the method filter used in the WebContent process
so that we will retain consistent protections when activating the GPU Process. These are the same rules
governing access to this resource when the GPU process is disabled, so should not create a change in behavior.

We also remove unneeded telemetry to reduce messaging burden and avoid performance issues.

* Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.sb:
* Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.sb:


Canonical link: https://commits.webkit.org/236971@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276515 268f45cc-cd09-0410-ab3c-d52691b4dbfc