REGRESSION (r273154): fast/forms/ios/repeatedly-focus-offscreen-select.html is consistently failing https://bugs.webkit.org/show_bug.cgi?id=224985 <rdar://problem/77042177>
Reviewed by Wenson Hsieh.

The test is failing after r273154, which made <select> elements present
a UIMenu rather than a popover. The old logic ensured popovers were not
presented when the <select> element was offscreen (see r265117 for more
information on why that behavior was necessary), but the new
presentation omitted that logic.

To achieve the correct behavior, and fix the failing test, <select>
menus should not be presented when the element is offscreen.

* UIProcess/ios/forms/WKFormSelectPicker.mm:
(-[WKSelectPicker controlBeginEditing]):


Canonical link: https://commits.webkit.org/236976@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276522 268f45cc-cd09-0410-ab3c-d52691b4dbfc