Fix interpolation of the clip CSS property https://bugs.webkit.org/show_bug.cgi?id=225017
Reviewed by Dean Jackson.

LayoutTests/imported/w3c:

Mark 13 WPT progressions.

* web-platform-tests/web-animations/animation-model/animation-types/accumulation-per-property-001-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/addition-per-property-001-expected.txt:
* web-platform-tests/web-animations/animation-model/animation-types/interpolation-per-property-001-expected.txt:
* web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-iteration-composite-operation-expected.txt:

Source/WebCore:

While we interpolated the clip value correctly, we wouldn't set the RenderStyle bit that
indicates that there is a non-auto value in the first place. With a new dedicated wrapper,
we can ensure we set the bit correctly as we blend, and that we use discrete interpolation
if both the from and to values aren't auto.

* animation/CSSPropertyAnimation.cpp:
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):


Canonical link: https://commits.webkit.org/236992@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276552 268f45cc-cd09-0410-ab3c-d52691b4dbfc