Crash in BreakBlockquoteCommand::doApply() https://bugs.webkit.org/show_bug.cgi?id=224941
Patch by Julian Gonzalez <julian_a_gonzalez@apple.com> on 2021-04-24
Reviewed by Ryosuke Niwa.

Source/WebCore:

Despite assertions to the contrary, it is possible for there not to be any node
to move into the new blockquote in BreakBlockquoteCommand::doApply() as a result
of layout updates, so remove the assertions and handle this case.

Test: editing/pasteboard/paste-as-quotation-then-paste-crash.html

* editing/BreakBlockquoteCommand.cpp:
(WebCore::BreakBlockquoteCommand::doApply):

LayoutTests:

Add test for this crash, running only on Release for now.
Thanks to Tuomas Karkkainen for its basic structure.

* TestExpectations:
* editing/pasteboard/paste-as-quotation-then-paste-crash-expected.txt: Added.
* editing/pasteboard/paste-as-quotation-then-paste-crash.html: Added.

Canonical link: https://commits.webkit.org/236995@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276558 268f45cc-cd09-0410-ab3c-d52691b4dbfc