Fix a crash caused by AVIF decoding failure https://bugs.webkit.org/show_bug.cgi?id=223843 <rdar://problem/76205084>
Reviewed by Philippe Normand.

Source/WebCore:

A crash happens when the image count is accessed in AVIFImageDecoder::tryDecodeSize.
This is because an instance of AVIFImageReader is destroyed by setFailed()
in AVIFImageReader::parseHeader. Deciding a failure was hasty because
parsing can fail when necessary data is not received yet. To fix this,
we call setFailed in AVIFImageReader::parseHeader only when all data is received.

Test: http/tests/images/avif-partial-load-crash.html

* platform/image-decoders/avif/AVIFImageDecoder.cpp:
(WebCore::AVIFImageDecoder::tryDecodeSize):
* platform/image-decoders/avif/AVIFImageReader.cpp:
(WebCore::AVIFImageReader::parseHeader):
(WebCore::AVIFImageReader::repetitionCount const):
* platform/image-decoders/avif/AVIFImageReader.h:

LayoutTests:

Check if partial load of avif images crashes.

* TestExpectations: Skip the test for non-gtk platforms.
* http/tests/images/avif-partial-load-crash-expected.txt: Added.
* http/tests/images/avif-partial-load-crash.html: Added.


Canonical link: https://commits.webkit.org/237014@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276578 268f45cc-cd09-0410-ab3c-d52691b4dbfc