ANGLE X11 build fixes https://bugs.webkit.org/show_bug.cgi?id=225003
Reviewed by Kenneth Russell.

X11 headers define a bunch of macros with common terms, interfering with WebCore and WTF
enum values. As a workaround, we explicitly undef them.

This is the same fix as the one to GLContext.h in r274869 but applied to the ANGLE headers
that are including X11.

* platform/graphics/angle/ANGLEHeaders.h:


Canonical link: https://commits.webkit.org/237027@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276591 268f45cc-cd09-0410-ab3c-d52691b4dbfc