Ask LocalStorage database to free as much memory as possible on memory pressure https://bugs.webkit.org/show_bug.cgi?id=225109
Reviewed by Geoffrey Garen.

Source/WebCore:

Add releaseMemory() function to SQLiteDatabase that calls sqlite3_db_release_memory()
to free as much memory as possible.

* platform/sql/SQLiteDatabase.cpp:
(WebCore::SQLiteDatabase::releaseMemory):
* platform/sql/SQLiteDatabase.h:

Source/WebKit:

Ask LocalStorage database to free as much memory as possible on memory pressure.
Some of the strings in the database can be large and we don't want SQLite to be
holding on to those for performance reason when we are under memory pressure.

This seems to be a ~1.75% improvement on PLUM/nytimes-home and a ~1% improvement
on PLUM/wikipedia, sadly not enough to show an impact on the overal PLUM score.

* NetworkProcess/NetworkProcess.cpp:
(WebKit::NetworkProcess::lowMemoryHandler):
* NetworkProcess/NetworkProcess.h:
* NetworkProcess/WebStorage/LocalStorageDatabase.cpp:
(WebKit::LocalStorageDatabase::handleLowMemoryWarning):
* NetworkProcess/WebStorage/LocalStorageDatabase.h:
* NetworkProcess/WebStorage/StorageArea.cpp:
(WebKit::StorageArea::handleLowMemoryWarning):
* NetworkProcess/WebStorage/StorageArea.h:
* NetworkProcess/WebStorage/StorageManagerSet.cpp:
(WebKit::StorageManagerSet::handleLowMemoryWarning):
* NetworkProcess/WebStorage/StorageManagerSet.h:


Canonical link: https://commits.webkit.org/237075@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276648 268f45cc-cd09-0410-ab3c-d52691b4dbfc